+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - inputs: **new** `keys` widget, for keyboard bindings
   - property inheritance: new special property name `_value` (retrieve the widget's value, and update when it changes)
   - switch: added `showValues` to display both values and labels in the buttons
+  - rgb: **removed** `touchAddress` option
 - misc
   - event system rewritten
   - js build size reduced